python/pyproject.toml
# https://docs.astral.sh/uv/concepts/projects/config/#configuring-projects
[project]
name = "readme-backend"
version = "1.0"
requires-python = ">=3.12"
dependencies = [
    #
    "fastapi",
    "uvicorn",
    "Pillow",
    "connect4.py==1.0.3",
    "minesweeper @ git+https://github.com/AiroPi/minesweeper",
]


[project.optional-dependencies]
debug = ["markdown"]

# https://docs.astral.sh/uv/concepts/projects/
[tool.uv]
dev-dependencies = [
    # "uv",
    "pyright",
    "ruff",
    "tox",
    "tox-uv",
    "types-Pillow",
]


# https://microsoft.github.io/pyright/#/
[tool.pyright]
typeCheckingMode = "strict"


# https://tox.wiki/en/latest/
[tool.tox]
envlist = ["py312"]

[tool.tox.env_run_base]
runner = "uv-venv-lock-runner"
description = "run all checks on the code"
extras = ["dev"]
commands = [
    #fmt
    ["ruff", "format", "--check", "src"],
    ["ruff", "check", "src"],
    ["pyright", "src"],
]


# https://docs.astral.sh/ruff/
[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"
src = ["src"]
exclude = ["bin"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle Error
    "F",     # pyflakes
    "UP",    # pyupgrade
    "SIM",   # flake8-simplify
    "I",     # imports
    "S",     # bandit (security)
    "N",     # pep8-naming
    "ASYNC", # flake8-async
    "C4",    # flake8-compehensions
    "FA",    # flake8-future-annotations
    "ISC",   # flake8-implicit-str-concat
    "ICN",   # flake8-import-conventions
    "G",     # flake8-logging-format
    "PIE",   # flake8-pie
    "PYI",   # flake8-pyi
    "RSE",   # flake8-raise
    "SLOT",  # flake8-slots
    "INT",   # flake8-gettext
    "TRY",   # tryceratops
    "FLY",   # flynt
    "PERF",  # Perflint
    "FURB",  # refurb
    "LOG",   # flake8-logging
    "RUF",   # Ruff-specific-rules
    # "ERA",   # locate commented codes
    # "FIX",   # locate TODOs and FIXME
    # "PTH",   # flake8-use-pathlib (maybe todo ?)
    # "TID",   # flake8-tidy-imports (maybe todo ?)
    # "SLF",   # flake8-self  (managed by pyright)
    # "RET",   # flake8-return
    # "Q",     # flake8-quotes
    # "T20",   # flake8-print
    # "DTZ",   # flake8-datetimez (TODO)
    # "B",     # flake8-bugbear
]

ignore = [
    "E501", # line too long (we relate on the formater)
    "N818", # Error suffix for exceptions names
    # "PIE796", # Enum contains duplicate value
    "TRY003", # Avoid specifying long messages outsides the exception class
    "ISC001", # To avoid conflicts with the formatter
]
dummy-variable-rgx = '^\*{0,2}(_$|__$|unused_|dummy_)'

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
combine-as-imports = true
